refactor(broker): split order_committer.rs into directory layout (C33 of PR-3)
Pure mechanical move. order_committer.rs (1059 lines) becomes:

- order_committer/mod.rs (~35 lines): module decls + pub(crate) use
  service::OrderCommitter + pub(crate) use types::{CommitmentComplete,
  CommitmentOutcome}.
- order_committer/error.rs (~36 lines): OrderCommitterErr enum +
  coded_error_impl!.
- order_committer/types.rs (~49 lines): CommitmentOutcome enum + its
  Display impl + CommitmentComplete (the capacity-release messages
  exchanged with downstream services).
- order_committer/state.rs (~36 lines): internal InFlightOrder and
  CommitterConfig state structs (both pub(super) so service.rs can
  construct them across modules).
- order_committer/service.rs (~973 lines): OrderCommitter struct + impl,
  BrokerService impl, the record_skip_telemetry helper, the three module
  constants (MIN_CAPACITY_CHECK_INTERVAL, MAX_PROVING_BATCH_SIZE,
  DEFAULT_MAX_COMMITMENT_DURATION_SECS), and the original test module.

External paths crate::order_committer::{OrderCommitter, CommitmentComplete,
CommitmentOutcome} preserved via the re-exports.

The CommitmentOutcome type itself is only constructed in tests (non-test
code uses CommitmentComplete.outcome via the Display impl), so the import
is moved into the tests block to avoid a non-test "unused import" warning.

The canonical plan's optional order_committer/prioritization.rs wrapper is
omitted: it would be pure indirection over the existing
crate::prioritization::prioritize_orders_to_commit shared helper.
